handle reportviewmodel, export PDF(incomplete)

+++ AirTicketSalesManagement/Models/ReportModel/MonthlyReportItem.cs
@@ -13,6 +13,6 @@ namespace AirTicketSalesManagement.Models.ReportModel
         public DateTime DepartureTime { get; set; }
         public int TicketsSold { get; set; }
         public decimal Revenue { get; set; }
-        public float RevenueRate { get; set; }
+        public decimal RevenueRate { get; set; }
     }
 }
